Reorganize Linodes store...
To be consistent with the new store pattern I split the Linode Duck
into reducer and actions. I've updated their export structure as well.
I then moved the events into the Linode store folder. All dependencies
have been updated to accomodate.
