+++ src/store/reducers/backupDrawer.ts
@@ -4,8 +4,8 @@ import { Reducer } from 'redux';
 import { ThunkAction } from 'redux-thunk';
 import { updateAccountSettings } from 'src/services/account';
 import { enableBackups } from 'src/services/linodes';
+import { updateMultipleLinodes } from 'src/store/linodes/linodes.actions';
 import { handleUpdate } from 'src/store/reducers/resources/accountSettings';
-import { updateMultipleLinodes } from 'src/store/reducers/resources/linodes';
 
 
 // HELPERS
